backend/requirements-prod.txt
fastapi>=0.115.0
uvicorn[standard]==0.29.0
langgraph>=1.0.10
anthropic>=0.40.0
qdrant-client==1.10.1
sentence-transformers==3.0.1
fastembed==0.3.6
assemblyai==0.59.0
PyJWT[crypto]>=2.12.0
httpx==0.27.0
pydantic-settings==2.2.1
structlog==24.1.0
slowapi==0.1.9
azure-storage-blob==12.20.0
